Split instance and spotRequest into separate classes
